@@ -1,9 +1,9 @@
 plugins {
-    id("com.github.johnrengelman.shadow") version "6.1.0"
-    id("io.micronaut.application") version "1.4.2"
-    id 'nu.studer.jooq' version '5.2'
-    id "org.flywaydb.flyway" version "7.7.2"
-    id 'com.avast.gradle.docker-compose' version '0.12.1'
+    id "com.github.johnrengelman.shadow" version "7.1.1"
+    id "io.micronaut.application" version "3.6.3"
+    id 'nu.studer.jooq' version '8.0'
+    id "org.flywaydb.flyway" version "9.7.0"
+    id 'com.avast.gradle.docker-compose' version '0.16.9'
 }
 
 version = "0.1"
@@ -104,29 +104,33 @@ dependencies {
     implementation("io.micronaut:micronaut-runtime")
     implementation("io.micronaut:micronaut-validation")
     runtimeOnly("ch.qos.logback:logback-classic")
-    annotationProcessor("io.micronaut.data:micronaut-data-processor:2.3.1")
+    annotationProcessor("io.micronaut.data:micronaut-data-processor:3.8.1")
+
+    implementation "io.reactivex.rxjava3:rxjava:3.1.5"
 
     // Lombok:
-    compileOnly 'org.projectlombok:lombok:1.18.16'
-    annotationProcessor 'org.projectlombok:lombok:1.18.16'
+    compileOnly 'org.projectlombok:lombok:1.18.24'
+    annotationProcessor 'org.projectlombok:lombok:1.18.24'
 
     // Micronaut inject
-    annotationProcessor(platform("io.micronaut:micronaut-bom:2.0.2"))
+    implementation(platform("io.micronaut:micronaut-bom:3.7.3"))
+    annotationProcessor(platform("io.micronaut:micronaut-bom:3.7.3"))
     annotationProcessor("io.micronaut:micronaut-inject-java")
-    implementation(platform("io.micronaut:micronaut-bom:2.0.2"))
     implementation("io.micronaut:micronaut-inject")
 
     // database configs
     implementation("io.micronaut.flyway:micronaut-flyway")
     implementation("io.micronaut.sql:micronaut-jdbc-hikari")
     implementation("io.micronaut.sql:micronaut-jooq")
-    jooqGenerator 'org.postgresql:postgresql:42.2.14'
-    implementation 'org.postgresql:postgresql:42.2.14'
+    jooqGenerator 'org.postgresql:postgresql:42.5.0'
+    implementation 'org.postgresql:postgresql:42.5.0'
+    implementation("jakarta.validation:jakarta.validation-api:3.0.2")
+    implementation("jakarta.persistence:jakarta.persistence-api:3.1.0")
 
     // security
     compileOnly "io.micronaut.security:micronaut-security-annotations"
     implementation "io.micronaut.security:micronaut-security-jwt"
-    implementation "org.springframework.security:spring-security-crypto:5.2.1.RELEASE"
+    implementation "org.springframework.security:spring-security-crypto:5.7.5"
 
     // transactional annotation
     implementation("io.micronaut.spring:micronaut-spring")
@@ -138,8 +142,8 @@ application {
 }
 
 java {
-    sourceCompatibility = JavaVersion.toVersion("1.8")
-    targetCompatibility = JavaVersion.toVersion("1.8")
+    sourceCompatibility = JavaVersion.toVersion("17")
+    targetCompatibility = JavaVersion.toVersion("17")
 }
 
 dockerCompose {
