@@ -3,13 +3,13 @@ package server.security;
 import io.micronaut.core.annotation.Nullable;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.security.authentication.*;
-import io.reactivex.BackpressureStrategy;
-import io.reactivex.Flowable;
+import io.reactivex.rxjava3.core.BackpressureStrategy;
+import io.reactivex.rxjava3.core.Flowable;
 import org.reactivestreams.Publisher;
 import server.account.repository.AccountRepository;
 
-import javax.inject.Inject;
-import javax.inject.Singleton;
+import jakarta.inject.Inject;
+import jakarta.inject.Singleton;
 
 @Singleton
 public class AuthenticationProviderUserPassword implements AuthenticationProvider {
@@ -20,6 +20,7 @@ public class AuthenticationProviderUserPassword implements AuthenticationProvide
     @Override
     public Publisher<AuthenticationResponse> authenticate(@Nullable HttpRequest<?> httpRequest,
                                                           AuthenticationRequest<?, ?> authenticationRequest) {
+
         return Flowable.create(emitter -> {
             String username = (String) authenticationRequest.getIdentity();
             String pw = (String) authenticationRequest.getSecret();
@@ -27,7 +28,7 @@ public class AuthenticationProviderUserPassword implements AuthenticationProvide
             boolean validCredentials = accountRepository.validCredentials(username, pw);
 
             if (validCredentials) {
-                emitter.onNext(new UserDetails(username, accountRepository.getRolesForUser(username)));
+                emitter.onNext(AuthenticationResponse.success(username, accountRepository.getRolesForUser(username)));
                 emitter.onComplete();
             } else {
                 emitter.onError(new AuthenticationException(new AuthenticationFailed()));
