@@ -1,11 +1,11 @@
 package server.security;
 
-import edu.umd.cs.findbugs.annotations.NonNull;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import javax.inject.Singleton;
+import jakarta.inject.Singleton;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 
 @Singleton
 public class BCryptPasswordEncoderService implements PasswordEncoder {
@@ -13,13 +13,13 @@ public class BCryptPasswordEncoderService implements PasswordEncoder {
     PasswordEncoder delegate = new BCryptPasswordEncoder();
 
     @Override
-    public String encode(@NotBlank @NonNull CharSequence rawPassword) {
+    public String encode(@NotBlank @NotNull CharSequence rawPassword) {
         return delegate.encode(rawPassword);
     }
 
     @Override
-    public boolean matches(@NotBlank @NonNull CharSequence rawPassword,
-                           @NotBlank @NonNull String encodedPassword) {
+    public boolean matches(@NotBlank @NotNull CharSequence rawPassword,
+                           @NotBlank @NotNull String encodedPassword) {
         return delegate.matches(rawPassword, encodedPassword);
     }
 }
